% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_sim_score.R
\name{calc_cluster_score}
\alias{calc_cluster_score}
\title{calculate the similarity score in a specific cluster.}
\usage{
calc_cluster_score(cluster, node1, node2)
}
\arguments{
\item{cluster,}{cluster as a list of indices of samples belonging to the
cluster.}

\item{node1,}{a nodes vector of net1}

\item{node2,}{a nodes vector of net2}
}
\description{
calculate the similarity score in a specific cluster.
}

