% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sim_score.R
\name{calc_sim_score}
\alias{calc_sim_score}
\title{Calculating the similarity scores}
\usage{
calc_sim_score(net, NFPnet, nperm = 100, ...)
}
\arguments{
\item{net, }{a graphNel object to represent the query biological networks,
for more details see \code{\link[graph]{graphNEL}}}

\item{NFPnet, }{a NFPRefnet object, one or more kegg pathway map, or
customized networks. For more details see \code{\link{NFPRefnet-class}}.}

\item{nperm, }{number of random networks for similarity score standardization}

\item{..., }{arguments passed to apcluster}
}
\value{
a similarity scoring vector,length is the same as the number of
networks
}
\description{
This function was used to calculate the similarity scores
between a network and the reference network.
}
\seealso{
\code{\link{NFPRefnet-class}}.
}
