% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_sim_score.R
\name{calc_twonet_score}
\alias{calc_twonet_score}
\title{calculate the similarity scores between two networks}
\usage{
calc_twonet_score(net1, net2, ...)
}
\arguments{
\item{net1,}{a igraph network}

\item{net2,}{same as net2}

\item{...,}{all other arguments are passed to apcluster}
}
\description{
calculate the similarity scores between two networks
}

