% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NFP-class.R
\docType{methods}
\name{cluster_info-methods}
\alias{cluster_info}
\alias{cluster_info,NFP-method}
\alias{cluster_info-methods}
\title{Extract the cluster information of \emph{NFP}.}
\usage{
cluster_info(object)

\S4method{cluster_info}{NFP}(object)
}
\arguments{
\item{object}{\code{NFP} object}
}
\value{
a list which contains the number, the examplar and some other cluster
properties.
}
\description{
This function extract the cluster information of network fingerprint.
}
\seealso{
\code{\link{NFP}}
}

