% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NFPRefnet-class.R
\docType{methods}
\name{group-methods}
\alias{group}
\alias{group,NFPRefnet-method}
\alias{group-methods}
\title{Group information of \emph{NFPRefnet}}
\usage{
group(object)

\S4method{group}{NFPRefnet}(object)
}
\arguments{
\item{object,}{\code{NFPRefnet} class}
}
\value{
a list which contains the group number and names of basic networks, as
well as the size of each group
}
\description{
This function extract the group information NFP basic networks.
}
\seealso{
\code{\link{NFPRefnet-class}}
}

