% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_data_package.R
\name{install_data_package}
\alias{install_data_package}
\title{Install NFP data package NFPdata}
\usage{
install_data_package(type = "ONL", loc = NULL)
}
\arguments{
\item{type}{A \code{string} with value \code{"ONL"} or \code{"LOCAL"}}

\item{loc}{A \code{string} that contains the file location.}
}
\description{
Downloads and Install the NFPdata Package to use with the NFP 
package
}
\details{
The NFPdata Package contains data that from kegg gene similarity
  based on gene ontology and is approximately a 16.4 MB download.
}
\examples{
\dontrun{
# Online install
install_data_package()
}
}
