% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kegg_refnet.R
\name{kegg_refnet}
\alias{kegg_refnet}
\title{Human KEGG signal pathway maps}
\format{A NFPRefnet object, more details see \code{\link{NFPRefnet-class}}}
\description{
A dataset containing the human signal pathway maps of KEGG
}
\seealso{
\code{\link{NFPRefnet-class}}
}

