% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadKEGGRefnet.R
\name{loadKEGGRefNet}
\alias{loadKEGGRefNet}
\title{Load the the reference molecular networks}
\usage{
loadKEGGRefNet(organism = "hsa")
}
\arguments{
\item{organsim,}{a character indicating to which organsim's pathway map was
taken as the basic network. e.g. \emph{hsa}.}
}
\value{
a KEGGRefnet object
}
\description{
This function generates the well-studied "basic networks".
}
\details{
KEGG pathway is a well-studied and the most widely used
biolocgial networks database. This function help users to load kegg pathway
maps as the basic networks.

Appanrently, users can also load their customozied biological networks as the
basic networks by creating a new \emph{NFPRefnet} object.
}
\seealso{
\code{\link{NFPRefnet-class}}
}

