% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NFPRefnet-class.R
\docType{methods}
\name{name-methods}
\alias{name}
\alias{name,NFPRefnet-method}
\alias{name-methods}
\title{Names of basic networks}
\usage{
name(object)

\S4method{name}{NFPRefnet}(object)
}
\arguments{
\item{object,}{\code{NFPRefnet} class}
}
\value{
a list
}
\description{
This function extract names of NFP basic networks.
}
\seealso{
\code{\link{NFPRefnet-class}}
}

