% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFPRefnet-class.R
\docType{methods}
\name{net-methods}
\alias{net}
\alias{net,NFPRefnet-method}
\alias{net-methods}
\title{Basic networks of \emph{NFPRefnet} class}
\usage{
net(object)

\S4method{net}{NFPRefnet}(object)
}
\arguments{
\item{object, }{\code{NFPRefnet} class}
}
\value{
a igraph list of all basic networks
}
\description{
This function extract the basic networks of NFPRefnet class.
}
\seealso{
\code{\link{NFPRefnet-class}}
}

