% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NFP-class.R
\docType{methods}
\name{perm_score-methods}
\alias{perm_score}
\alias{perm_score,NFP-method}
\alias{perm_score-methods}
\title{Extract the randomized similarity score}
\usage{
perm_score(object)

\S4method{perm_score}{NFP}(object)
}
\arguments{
\item{object,}{\code{NFP} class}
}
\value{
a data frame, each col (elements) represents once permutation
similarity score, each row indicate a reference basic network.
}
\description{
This function extract the randomized similarity score for standardization.
}
\seealso{
\code{\link{NFP}}
}

