% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NFP-class.R
\docType{methods}
\name{plot,NFP,ANY-method}
\alias{plot}
\alias{plot,NFP,ANY-method}
\alias{plot-methods}
\title{Plot NFP results}
\usage{
\S4method{plot}{NFP,ANY}(x, type = c("matchstick", "line", "point"),
  p_size = 3, l_size = 1)
}
\arguments{
\item{x,}{\code{NFP} class}

\item{type,}{types of the visaulization of \emph{NFP} object, point or line.
Default is point.}

\item{p_size,}{point size of plot, default is 3.}

\item{l_size,}{line size of plot, default is 1.}
}
\description{
Function for visualization NFP results.
}
\seealso{
\code{\link{NFP-class}}
}

