% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFP-class.R
\docType{methods}
\name{plot_NFP-methods}
\alias{plot_NFP-methods}
\alias{plot_NFP,NFP-method}
\alias{plot_NFP}
\title{Plot NFP results}
\usage{
plot_NFP(object, type = c("matchstick", "line", "point"), p_size = 2,
  l_size = 0.5)

\S4method{plot_NFP}{NFP}(object, type = c("matchstick", "line", "point"),
  p_size = 2, l_size = 0.5)
}
\arguments{
\item{object, }{\code{NFP} class}

\item{type, }{types of the visaulization of \emph{NFP} object, point or line.
Default is point.}

\item{p_size, }{point size of plot, default is 2.}

\item{l_size, }{line size of plot, default is 0.5.
#'@aliases plot_NFP plot_NFP-methods}
}
\description{
Function for visualization NFP results.
}
\seealso{
\code{\link{NFP-class}}
}
