% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NFPlist.R
\name{plot_NFPlist}
\alias{plot_NFPlist}
\title{Plot multiple NFPs.}
\usage{
plot_NFPlist(object, l_size = 0.5)
}
\arguments{
\item{object, }{\code{NFP} class list.}

\item{l_size, }{line size of plot, default is 0.5.}
}
\description{
Function for visualization multiple NFPs.
}
\seealso{
\code{\link{NFP-class}}
}
