% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NFPRefnet-class.R
\docType{methods}
\name{refnet_name-methods}
\alias{refnet_name}
\alias{refnet_name,NFPRefnet-method}
\alias{refnet_name-methods}
\title{Names of basic networks}
\usage{
refnet_name(object)

\S4method{refnet_name}{NFPRefnet}(object)
}
\arguments{
\item{object,}{\code{NFPRefnet} class}
}
\value{
a list
}
\description{
This function extract names of NFP basic networks.
}
\seealso{
\code{\link{NFPRefnet-class}}
}

