% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFP-class.R
\docType{methods}
\name{show,NFP-method}
\alias{show}
\alias{show,NFP-method}
\alias{show-methods}
\title{The show generic function}
\usage{
\S4method{show}{NFP}(object)
}
\arguments{
\item{object, }{\code{NFP} object}
}
\description{
Show a shor summary for NFP object, see \code{\link[methods]{show}}.
}

