% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NFP-class.R
\docType{methods}
\name{show,NFP-method}
\alias{show,NFP-method}
\alias{show_NFP}
\alias{show_NFP-methods}
\title{Show an Object}
\usage{
\S4method{show}{NFP}(object)
}
\arguments{
\item{object,}{\code{NFP} object}
}
\description{
show method short for NFP object, see \code{\link[methods]{show}}.
}

