% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NFPRefnet-class.R
\docType{methods}
\name{show,NFPRefnet-method}
\alias{show,NFPRefnet-method}
\alias{show_NFPRefnet}
\alias{show_NFPRefnet-methods}
\title{Show an Object}
\usage{
\S4method{show}{NFPRefnet}(object)
}
\arguments{
\item{object,}{\code{NFPRefnet} class}
}
\description{
show method short for NFPRefnet object, see \code{\link[methods]{show}}
}

