% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_sim_score.R
\name{standardize_score}
\alias{standardize_score}
\title{standardize the similarity score}
\usage{
standardize_score(net, NFPnet, nperm, ...)
}
\arguments{
\item{net,}{igraph object}

\item{NFPnet,}{NFPRefnet, the reference NFP networks}

\item{nperm,}{permutation number for NFP standardization}

\item{...,}{argument passed to calc_twonet_score}
}
\description{
standardize the similarity score
}

