% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFP-class.R
\docType{methods}
\name{sub_NFP-methods}
\alias{sub_NFP}
\alias{sub_NFP,NFP-method}
\alias{sub_NFP-methods}
\title{subset of NFP object}
\usage{
sub_NFP(object, i)

\S4method{sub_NFP}{NFP}(object, i)
}
\arguments{
\item{object, }{\code{NFP} class}

\item{i, }{numeric or character indicating the index or the names of the
reference network}
}
\value{
an similar NFP object contain just the selected elements.
}
\description{
This function extract the subsets of NFP-class.
}
\seealso{
\code{\link{NFP-class}}
}

