% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NFPRefnet-class.R
\docType{methods}
\name{subnet-methods}
\alias{subnet}
\alias{subnet,NFPRefnet-method}
\alias{subnet-methods}
\title{Subset the basic networks}
\usage{
subnet(object, group_name, index = NULL)

\S4method{subnet}{NFPRefnet}(object, group_name, index = NULL)
}
\arguments{
\item{object,}{\code{NFPRefnet} class.}

\item{group_name,}{character, indicating the groups to subset.}

\item{index,}{numeric, character or NA, indices specifying elements to extract. This
parameter only works while \code{group_name} is a length-one character.
Default is \emph{NULL}, indicating extract all the networks of a group. See
\emph{details} for more information.}
}
\description{
Extract or Replace parts of the NFP basic networks.
}
\details{
This function help users to extract the specific networks for
customized analysis, which could be of entire group networks or some part of
a specific group networks.subsequent analysis.

Note, the \code{index} argument is only worked while one argument is
consideration, which means group_name is a length-one character. And default
is \emph{NULL}, indicating extract the entire group basic networks.
}
\seealso{
\code{\link{NFPRefnet-class}}
}

