% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_sim_score.R
\name{calc_sim_score}
\alias{calc_sim_score}
\title{Calculating the similarity scores}
\usage{
calc_sim_score(net, refnet, nperm = 100, plot = TRUE, ...)
}
\arguments{
\item{net,}{a network, can be represented as adj matrix, edge list, or in
igraph class}

\item{refnet,}{a reference network list, one or more kegg pathway map, or
customized networks, can be represented as adj matrix, edge list, or in
igraph class}

\item{nperm,}{number of random networks for similarity score standardization}
}
\value{
a similarity scoring vector,length is the same as the number of
networks
}
\description{
This function was used to calculate the similarity scores
between a network and the reference network.
}

