% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadKEGGRefNet.R
\name{loadKEGGRefNet}
\alias{loadKEGGRefNet}
\title{Load the kegg signal pathway map as the reference net}
\usage{
loadKEGGRefNet()
}
\value{
a igraph list which represents the kegg signal pathway
}
\description{
This function generates the well-studied KEGG signal pathway
which was taken as the reference net.
}

